Исправлены ошибки

+++ telegram-bot/src/main/java/pro/sky/telegrambot/repository/UserRepository.java
@@ -19,7 +19,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findById(UUID telegramId);
     Optional<User> getUserByChatId(long chatId);
 
-    User findUserByChatId(int chatId);
+    User findUserByChatId(long chatId);
 
 //    @Query("SELECT CASE WHEN COUNT(u) > 0 THEN true ELSE false END FROM User u WHERE u.chatId = :chatId")
 
@@ -32,5 +32,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
     @Transactional
     @Modifying
     @Query("UPDATE User u set u.number = ?2 where u.chatId = ?1")
-    int updateNumber(int chatId, String num);
+    int updateNumber(long chatId, String num);
 }
